[exec][cli] Move common options to klio core from exec and cli
