+++ cli/src/klio_cli/options.py
@@ -15,56 +15,7 @@
 
 import click
 
-
-#####
-# utils for options
-#####
-class MutuallyExclusiveOption(click.Option):
-    """
-    Helper class to validate and document mutual exclusivity between options.
-
-    This unfortunately doesn't work with click arguments (only options).
-
-    To use, pass in both of the following keywords into an option declaration:
-        click.option(
-            "--an-option"
-            cls=MutuallyExclusiveOption
-            mutually_exclusive=["string_of_exclusive_option", "another_option"]
-        )
-    """
-
-    def __init__(self, *args, **kwargs):
-        self.mut_ex_opts = set(kwargs.pop("mutually_exclusive", []))
-        help_text = kwargs.get("help", "")
-        if self.mut_ex_opts:
-            mutex = [self._varname_to_opt_flag(m) for m in self.mut_ex_opts]
-            mutex_fmted = ["``{}``".format(m) for m in mutex]
-            ex_str = ", ".join(mutex_fmted)
-            kwargs["help"] = help_text + (
-                "\n\n**NOTE:** This option is mutually exclusive with "
-                "[{}].".format(ex_str)
-            )
-        super(MutuallyExclusiveOption, self).__init__(*args, **kwargs)
-
-    @staticmethod
-    def _varname_to_opt_flag(var):
-        return "--" + var.replace("_", "-")
-
-    def handle_parse_result(self, ctx, opts, args):
-        if self.mut_ex_opts.intersection(opts) and self.name in opts:
-            mutex = [
-                "`" + self._varname_to_opt_flag(m) + "`"
-                for m in self.mut_ex_opts
-            ]
-            mutex = ", ".join(mutex)
-            msg = "Illegal usage: `{}` is mutually exclusive with {}.".format(
-                self._varname_to_opt_flag(self.name), mutex
-            )
-            raise click.UsageError(msg)
-
-        return super(MutuallyExclusiveOption, self).handle_parse_result(
-            ctx, opts, args
-        )
+import klio_core.options as core_options
 
 
 def _verify_gcs_uri(ctx, param, value):
@@ -92,7 +43,7 @@ def job_dir(*args, **kwargs):
                 "Job directory where the job's ``Dockerfile`` is located. "
                 "Defaults current working directory."
             ),
-            cls=MutuallyExclusiveOption,
+            cls=core_options.MutuallyExclusiveOption,
             mutually_exclusive=mutually_exclusive,
         )(func)
 
@@ -116,7 +67,7 @@ def config_file(*args, **kwargs):
                 "will be treated relative to ``--job-dir``. Defaults to "
                 "``klio-job.yaml``."
             ),
-            cls=MutuallyExclusiveOption,
+            cls=core_options.MutuallyExclusiveOption,
             mutually_exclusive=mutually_exclusive,
         )(func)
 
@@ -151,15 +102,6 @@ def template(func):
     )(func)
 
 
-def image_tag(func):
-    return click.option(
-        "--image-tag",
-        default=None,
-        show_default="``git-sha[dirty?]``",
-        help="Docker image tag to use.",
-    )(func)
-
-
 def force_build(func):
     return click.option(
         "--force-build",
@@ -169,28 +111,10 @@ def force_build(func):
     )(func)
 
 
-def direct_runner(func):
-    return click.option(
-        "--direct-runner",
-        default=False,
-        is_flag=True,
-        help="Run the job locally via the DirectRunner.",
-    )(func)
-
-
-def update(func):
-    return click.option(
-        "--update/--no-update",
-        default=None,
-        is_flag=True,
-        help="[Experimental] Update an existing streaming Cloud Dataflow job.",
-    )(func)
-
-
 def runtime(func):
     func = force_build(func)
-    func = direct_runner(func)
-    func = update(func)
+    func = core_options.direct_runner(func)
+    func = core_options.update(func)
     return func
 
 
@@ -284,7 +208,7 @@ def input_file(*args, **kwargs):
             "-i",
             "--input-file",
             type=click.Path(exists=False, dir_okay=False, readable=True),
-            cls=MutuallyExclusiveOption,
+            cls=core_options.MutuallyExclusiveOption,
             mutually_exclusive=mutually_exclusive,
             **kwargs,
         )(func)
@@ -304,7 +228,7 @@ def output_file(*args, **kwargs):
             "-o",
             "--output-file",
             type=click.Path(exists=False, dir_okay=False, writable=True),
-            cls=MutuallyExclusiveOption,
+            cls=core_options.MutuallyExclusiveOption,
             mutually_exclusive=mutually_exclusive,
             **kwargs,
         )(func)
@@ -316,16 +240,6 @@ def output_file(*args, **kwargs):
     return wrapper
 
 
-def show_logs(func):
-    return click.option(
-        "--show-logs",
-        default=False,
-        show_default=True,
-        is_flag=True,
-        help="Show a job's logs while profiling.",
-    )(func)
-
-
 #####
 # options for `klio job profile collect-profiling-data`
 #####
@@ -335,7 +249,7 @@ def gcs_location(func):
         default=None,
         show_default="`pipeline_options.profile_location` in `klio-job.yaml`",
         help="GCS location of cProfile data.",
-        cls=MutuallyExclusiveOption,
+        cls=core_options.MutuallyExclusiveOption,
         mutually_exclusive=["job_dir", "input_file", "config_file"],
         callback=_verify_gcs_uri,
     )(func)
@@ -398,101 +312,6 @@ def sort_stats(func):
             "pstats.Stats.sort_stats>`_. Multiple ``--sort-stats`` invocations "
             "are supported."
         ),
-    )(func)
-
-
-#####
-# options for `klio job profile memory`
-#####
-def interval(func):
-    return click.option(
-        "--interval",
-        default=0.1,
-        show_default=True,
-        type=float,
-        help="Sampling period (in seconds).",
-    )(func)
-
-
-def include_children(func):
-    return click.option(
-        "--include-children",
-        default=False,
-        show_default=True,
-        is_flag=True,
-        help="Monitor forked processes as well (sums up all process memory).",
-    )(func)
-
-
-def multiprocess(func):
-    return click.option(
-        "--multiprocess",
-        default=False,
-        show_default=True,
-        is_flag=True,
-        help=(
-            "Monitor forked processes creating individual plots for each "
-            "child."
-        ),
-    )(func)
-
-
-def plot_graph(func):
-    return click.option(
-        "-g",
-        "--plot-graph",
-        default=False,
-        show_default=True,
-        is_flag=True,
-        help=(
-            "Plot memory profile using matplotlib. Saves to "
-            "``klio_profile_memory_<YYYYMMDDhhmmss>.png``."
-        ),
-    )(func)
-
-
-#####
-# options for `klio job profile memory-per-line`
-#####
-def maximum(func):
-    return click.option(
-        "--maximum",
-        "get_maximum",
-        default=False,
-        show_default=True,
-        is_flag=True,
-        cls=MutuallyExclusiveOption,
-        mutually_exclusive=["per_element"],
-        help=(
-            "Print maximum memory usage per line in aggregate of all input "
-            "elements process."
-        ),
-    )(func)
-
-
-def per_element(func):
-    return click.option(
-        "--per-element",
-        default=False,
-        show_default=True,
-        is_flag=True,
-        cls=MutuallyExclusiveOption,
-        mutually_exclusive=["maximum"],
-        help="Print memory usage per line for each input element processed.",
-    )(func)
-
-
-#####
-# options for `klio job profile timeit`
-#####
-def iterations(func):
-    return click.option(
-        "-n",
-        "--iterations",
-        default=10,
-        show_default=True,
-        type=int,
-        help="Number of times to execute each entity ID provided.",
     )(func)
