+++ core/setup.py
@@ -151,6 +151,7 @@ PROJECT_URLS = {
 }
 META_FILE = read(META_PATH)
 INSTALL_REQUIRES = [
+    "click",
     "glom",
     "google-api-python-client>=1.10.0,<1.12",
     "google-api-core>1.18.0,<1.21.0",  # TODO: try and remove
