+++ docs/src/reference/core/api/options.rst
@@ -1,0 +1,18 @@
+``klio_core.options``
+===================
+
+.. automodule:: klio_core.options
+
+.. currentmodule:: klio_core.options
+
+.. autofunction:: image_tag
+.. autofunction:: direct_runner
+.. autofunction:: update
+.. autofunction:: show_logs
+.. autofunction:: interval
+.. autofunction:: include_children
+.. autofunction:: multiprocess
+.. autofunction:: plot_graph
+.. autofunction:: maximum
+.. autofunction:: per_element
+.. autofunction:: iterations
